requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
passlib[bcrypt]
python-jose[cryptography]
python-multipart
httpx
qdrant-client
sentence-transformers
python-dotenv
streamlit
pandas
# Cloud LLM Providers
openai>=1.0.0
google-generativeai>=0.3.0
edge-tts
google-cloud-texttospeech
groq
mss
numpy
PyQt6
pyaudio
pvporcupine
sounddevice
vosk
websockets
pygetwindow
